05-Auth-pt1/server/config.py
from dotenv import load_dotenv
load_dotenv()

import os

from flask import Flask
from flask_bcrypt import Bcrypt
from flask_migrate import Migrate
from flask_restful import Api
from flask_sqlalchemy import SQLAlchemy

app = Flask(
    __name__,
    static_url_path="",
    static_folder="../client/build",
    template_folder="../client/build"
    )
app.config['SQLALCHEMY_DATABASE_URI'] = os.environ.get('DATABASE_URI')
app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
app.json.compact = False

app.secret_key = b'@~xH\xf2\x10k\x07hp\x85\xa6N\xde\xd4\xcd'

db = SQLAlchemy()
migrate = Migrate(app, db)

db.init_app(app)

bcrypt = Bcrypt(app)

# monkeypatch flask-rest Api class to bypass its error routing
Api.error_router = lambda self, hnd, e: hnd(e)
api = Api(app)