05-Auth-pt1/requirements.txt
-i https://pypi.org/simple
alembic==1.10.4; python_version >= '3.7'
aniso8601==9.0.1
appnope==0.1.3; sys_platform == 'darwin'
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.0.1; python_version >= '3.6'
blinker==1.6.2; python_version >= '3.7'
click==8.1.3; python_version >= '3.7'
decorator==5.1.1; python_version < '3.11' and python_version >= '3.7'
executing==1.2.0
faker==18.7.0
flask==2.3.2
flask-bcrypt==1.0.1
flask-cors==3.0.10
flask-migrate==4.0.4
flask-restful==0.3.9
flask-sqlalchemy==3.0.3
gunicorn==20.1.0
honcho==1.1.0
importlib-metadata==6.6.0; python_version < '3.10'
importlib-resources==5.12.0
ipdb==0.13.13
ipython==8.12.2; python_version < '3.11' and python_version >= '3.7'
itsdangerous==2.1.2; python_version >= '3.7'
jedi==0.18.2; python_version >= '3.6'
jinja2==3.1.2; python_version >= '3.7'
jsonify==0.5
mako==1.2.4; python_version >= '3.7'
markupsafe==2.1.2; python_version >= '3.7'
matplotlib-inline==0.1.6; python_version >= '3.5'
parso==0.8.3; python_version >= '3.6'
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
prompt-toolkit==3.0.38; python_version >= '3.7'
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.15.1; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
pytz==2023.3
setuptools==67.7.2; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.13; python_version >= '3.7'
sqlalchemy-serializer==1.4.1
stack-data==0.6.2
tomli==2.0.1; python_version < '3.11' and python_version >= '3.7'
traitlets==5.9.0; python_version >= '3.7'
typing-extensions==4.5.0; python_version >= '3.7'
wcwidth==0.2.6
werkzeug==2.3.4; python_version >= '3.8'
zipp==3.15.0; python_version < '3.10'
